+++ src/main/java/com/turkraft/springfilter/DocumentExecutor.java
@@ -4,20 +4,42 @@ import java.util.List;
 import java.util.Optional;
 import org.bson.Document;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.data.mongodb.core.MongoOperations;
+import org.springframework.data.mongodb.repository.query.MongoEntityInformation;
 import org.springframework.lang.Nullable;
 
-public interface DocumentExecutor<T> {
+public interface DocumentExecutor<T, I> {
 
-  Optional<T> findOne(@Nullable Document doc);
+  default Optional<T> findOne(@Nullable Document doc) {
+    return Optional.ofNullable(getMongoOperations().findOne(FilterUtils.getQueryFromDocument(doc),
+        getMetadata().getJavaType()));
+  }
 
-  List<T> findAll(@Nullable Document doc);
+  default List<T> findAll(@Nullable Document doc) {
+    return getMongoOperations().find(FilterUtils.getQueryFromDocument(doc),
+        getMetadata().getJavaType());
+  }
 
-  Page<T> findAll(@Nullable Document doc, Pageable pageable);
+  default Page<T> findAll(@Nullable Document doc, Pageable pageable) {
+    return new PageImpl<>(getMongoOperations()
+        .find(FilterUtils.getQueryFromDocument(doc).with(pageable), getMetadata().getJavaType()));
+  }
 
-  List<T> findAll(@Nullable Document doc, Sort sort);
+  default List<T> findAll(@Nullable Document doc, Sort sort) {
+    return getMongoOperations().find(FilterUtils.getQueryFromDocument(doc).with(sort),
+        getMetadata().getJavaType());
+  }
 
-  long count(@Nullable Document doc);
+  default long count(@Nullable Document doc) {
+    return getMongoOperations().count(FilterUtils.getQueryFromDocument(doc),
+        getMetadata().getJavaType());
+  }
+
+  MongoEntityInformation<T, I> getMetadata();
+
+  MongoOperations getMongoOperations();
 
 }
